requirement.txt
django==3.2.6
django-ninja==0.14.0